Put app code in ./App folder
