+++ app.json
@@ -4,7 +4,8 @@
     "slug": "bible",
     "version": "1.0.0",
     "orientation": "portrait",
-    "icon": "./assets/icon.png",
+		"icon": "./assets/icon.png",
+		"entryPoint": "./App/App.js",
     "splash": {
       "image": "./assets/splash.png",
       "resizeMode": "contain",
@@ -21,6 +22,7 @@
     },
     "web": {
       "favicon": "./assets/favicon.png"
-    }
+    },
+    "description": ""
   }
 }
